Relocate Collection to pants.util.objects for broader use.
