+++ src/python/pants/engine/fs.py
@@ -9,10 +9,9 @@ from binascii import hexlify
 from os.path import join
 
 from pants.base.project_tree import Dir, File
-from pants.engine.addressable import Collection
 from pants.engine.rules import RootRule, rule
 from pants.engine.selectors import Select
-from pants.util.objects import datatype
+from pants.util.objects import Collection, datatype
 
 
 class FileContent(datatype('FileContent', ['path', 'content'])):
